v0.2.1: Add system registers, multi-platform support, and licensing
- Add ds_system_register_names.h: library-owned system registers
  (DS_PACKET_COUNT, DS_ERROR_COUNT, CONTROL_INTERFACE, COUNTER_1HZ)
  accessed via SYS_REGS, separate from user registers (REGS)
- Update dsRegisters.h and dsRegisters.c to handle system register
  read/write via dedicated packet types
- Expand README with ESP-IDF and STM32CubeIDE integration guides
- Rework USER_INTEGRATION_GUIDE.md for clarity and completeness
- Update ds_minimal_register_names.h and ds_minimal_parameter_names.h
- Update example register names header to reflect system register split
- Add LICENSE (PolyForm Noncommercial) and COMMERCIAL.md
- Bump version to 0.2.1

+++ src/dsTCP.c
@@ -321,7 +321,7 @@ void prvdsTCPTask(void *pvParameters)
             else if (lBytes >= 0)
             {
                 // Packet size error, get proper response from datastream
-                dsPacketSizeError(lBytes, &txPacket);
+                dsPacketSizeError((uint32_t)lBytes, &txPacket);
 
                 lSent = dsTCPSend(connectedSocket, txPacket.buffer, DATASTREAM_OUTPUT_SIZE, 0);
